third_party/blink/web_tests/platform/win/external/wpt/url/a-element-xhtml_exclude=(file_javascript_mailto)-expected.txt

This is a testharness.js-based test.
Found 49 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Parsing: <a:	 foo.com> against <http://example.org/foo/bar>
  assert_equals: href expected "a: foo.com" but got "file:///A:/%20foo.com"
[FAIL] Parsing: <\\\\x\\hello> against <http://example.org/foo/bar>
  assert_equals: href expected "http://x/hello" but got "file://x/hello"
[FAIL] Parsing: <c:/foo> against <http://example.org/foo/bar>
  assert_equals: href expected "c:/foo" but got "file:///C:/foo"
[FAIL] Parsing: <http://example.com/foo/%2e%2> against <about:blank>
  assert_equals: href expected "http://example.com/foo/%2e%2" but got "http://example.com/foo/.%2"
[FAIL] Parsing: <http://example.com/foo/%2e./%2e%2e/.%2e/%2e.bar> against <about:blank>
  assert_equals: href expected "http://example.com/%2e.bar" but got "http://example.com/..bar"
[FAIL] Parsing: <http://www/foo%2Ehtml> against <about:blank>
  assert_equals: href expected "http://www/foo%2Ehtml" but got "http://www/foo.html"
[FAIL] Parsing: <http://example example.com> against <http://other.com/>
  assert_unreached: Expected URL to fail parsing Reached unreachable code
[FAIL] Parsing: <http://GOO 　goo.com> against <http://other.com/>
  assert_unreached: Expected URL to fail parsing Reached unreachable code
[FAIL] Parsing: <http://a.b.c.xn--pokxncvks> against <about:blank>
  assert_unreached: Expected URL to fail parsing Reached unreachable code
[FAIL] Parsing: <http://10.0.0.xn--pokxncvks> against <about:blank>
  assert_unreached: Expected URL to fail parsing Reached unreachable code
[FAIL] Parsing: <http://a.b.c.XN--pokxncvks> against <about:blank>
  assert_unreached: Expected URL to fail parsing Reached unreachable code
[FAIL] Parsing: <http://a.b.c.Xn--pokxncvks> against <about:blank>
  assert_unreached: Expected URL to fail parsing Reached unreachable code
[FAIL] Parsing: <http://10.0.0.XN--pokxncvks> against <about:blank>
  assert_unreached: Expected URL to fail parsing Reached unreachable code
[FAIL] Parsing: <http://10.0.0.xN--pokxncvks> against <about:blank>
  assert_unreached: Expected URL to fail parsing Reached unreachable code
[FAIL] Parsing: <http://％４１.com> against <http://other.com/>
  assert_unreached: Expected URL to fail parsing Reached unreachable code
[FAIL] Parsing: <http://%ef%bc%85%ef%bc%94%ef%bc%91.com> against <http://other.com/>
  assert_unreached: Expected URL to fail parsing Reached unreachable code
[FAIL] Parsing: <http://192.168.0.1 hello> against <http://other.com/>
  assert_unreached: Expected URL to fail parsing Reached unreachable code
[FAIL] Parsing: <https://x x:12> against <about:blank>
  assert_unreached: Expected URL to fail parsing Reached unreachable code
[FAIL] Parsing: <h://.> against <about:blank>
  assert_equals: href expected "h://." but got "file:///H://"
[FAIL] Parsing: <http://[::%31]> against <http://other.com/>
  assert_unreached: Expected URL to fail parsing Reached unreachable code
[FAIL] Parsing: <http://%5B::1]> against <http://other.com/>
  assert_unreached: Expected URL to fail parsing Reached unreachable code
[FAIL] Parsing: <notspecial://host/?'> against <about:blank>
  assert_equals: href expected "notspecial://host/?'" but got "notspecial://host/?%27"
[FAIL] Parsing: <http://a b/> against <about:blank>
  assert_unreached: Expected URL to fail parsing Reached unreachable code
[FAIL] Parsing: <http://ho%20st/> against <about:blank>
  assert_unreached: Expected URL to fail parsing Reached unreachable code
[FAIL] Parsing: <http://!"$&'()*+,-.;=_`{}~/> against <about:blank>
  assert_equals: href expected "http://!\\"$&'()*+,-.;=_`{}~/" but got "http://!\\"$&'()%2A+,-.;=_`{}~/"
[FAIL] Parsing: <\\\\\\.\\Y:> against <about:blank>
  assert_unreached: Expected URL to fail parsing Reached unreachable code
[FAIL] Parsing: <\\\\\\.\\y:> against <about:blank>
  assert_unreached: Expected URL to fail parsing Reached unreachable code
[FAIL] Parsing: <test-a-colon.html> against <a:>
  assert_unreached: Expected URL to fail parsing Reached unreachable code
[FAIL] Parsing: <test-a-colon-b.html> against <a:b>
  assert_unreached: Expected URL to fail parsing Reached unreachable code
[FAIL] Parsing: <test-a-colon-slash.html> against <a:/>
  assert_equals: href expected "a:/test-a-colon-slash.html" but got "file:///A:/test-a-colon-slash.html"
[FAIL] Parsing: <test-a-colon-slash-slash.html> against <a://>
  assert_equals: href expected "a:///test-a-colon-slash-slash.html" but got "file:///A:/.//test-a-colon-slash-slash.html"
[FAIL] Parsing: <test-a-colon-slash-b.html> against <a:/b>
  assert_equals: href expected "a:/test-a-colon-slash-b.html" but got "file:///A:/test-a-colon-slash-b.html"
[FAIL] Parsing: <test-a-colon-slash-slash-b.html> against <a://b>
  assert_equals: href expected "a://b/test-a-colon-slash-slash-b.html" but got "file:///A:/.//test-a-colon-slash-slash-b.html"
[FAIL] Parsing: <foo:// !"$%&'()*+,-.;<=>@[\\]^_`{|}~@host/> against <about:blank>
  assert_equals: href expected "foo://%20!%22$%&'()*+,-.%3B%3C%3D%3E%40%5B%5C%5D%5E_%60%7B%7C%7D~@host/" but got "foo://%20!%22$%&%27()*+,-.%3B%3C%3D%3E%40%5B%5C%5D%5E_%60%7B%7C%7D~@host/"
[FAIL] Parsing: <wss:// !"$%&'()*+,-.;<=>@[]^_`{|}~@host/> against <about:blank>
  assert_equals: href expected "wss://%20!%22$%&'()*+,-.%3B%3C%3D%3E%40%5B%5D%5E_%60%7B%7C%7D~@host/" but got "wss://%20!%22$%&%27()*+,-.%3B%3C%3D%3E%40%5B%5D%5E_%60%7B%7C%7D~@host/"
[FAIL] Parsing: <foo://joe: !"$%&'()*+,-.:;<=>@[\\]^_`{|}~@host/> against <about:blank>
  assert_equals: href expected "foo://joe:%20!%22$%&'()*+,-.%3A%3B%3C%3D%3E%40%5B%5C%5D%5E_%60%7B%7C%7D~@host/" but got "foo://joe:%20!%22$%&%27()*+,-.%3A%3B%3C%3D%3E%40%5B%5C%5D%5E_%60%7B%7C%7D~@host/"
[FAIL] Parsing: <wss://joe: !"$%&'()*+,-.:;<=>@[]^_`{|}~@host/> against <about:blank>
  assert_equals: href expected "wss://joe:%20!%22$%&'()*+,-.%3A%3B%3C%3D%3E%40%5B%5D%5E_%60%7B%7C%7D~@host/" but got "wss://joe:%20!%22$%&%27()*+,-.%3A%3B%3C%3D%3E%40%5B%5D%5E_%60%7B%7C%7D~@host/"
[FAIL] Parsing: <wss://!"$&'()*+,-.;=_`{}~/> against <about:blank>
  assert_equals: href expected "wss://!\\"$&'()*+,-.;=_`{}~/" but got "wss://!\\"$&'()%2A+,-.;=_`{}~/"
[FAIL] Parsing: <foo://host/ !"$%&'()*+,-./:;<=>@[\\]^_`{|}~> against <about:blank>
  assert_equals: href expected "foo://host/%20!%22$%&'()*+,-./:;%3C=%3E@[\\\\]^_%60%7B|%7D~" but got "foo://host/%20!%22$%&'()*+,-./:;%3C=%3E@[\\\\]%5E_%60%7B%7C%7D~"
[FAIL] Parsing: <wss://host/ !"$%&'()*+,-./:;<=>@[\\]^_`{|}~> against <about:blank>
  assert_equals: href expected "wss://host/%20!%22$%&'()*+,-./:;%3C=%3E@[/]^_%60%7B|%7D~" but got "wss://host/%20!%22$%&'()*+,-./:;%3C=%3E@[/]%5E_%60%7B%7C%7D~"
[FAIL] Parsing: <foo://host/dir/? !"$%&'()*+,-./:;<=>?@[\\]^_`{|}~> against <about:blank>
  assert_equals: href expected "foo://host/dir/?%20!%22$%&'()*+,-./:;%3C=%3E?@[\\\\]^_`{|}~" but got "foo://host/dir/?%20!%22$%&%27()*+,-./:;%3C=%3E?@[\\\\]^_`{|}~"
[FAIL] Parsing: <https://xn--/> against <about:blank>
  assert_unreached: Expected URL to fail parsing Reached unreachable code
[FAIL] Parsing: <w://x:0> against <about:blank>
  assert_equals: href expected "w://x:0" but got "file:///W://x:0"
[FAIL] Parsing: <android://x:0/a> against <about:blank>
  assert_equals: host expected "x:0" but got ""
[FAIL] Parsing: <drivefs://x:0/a> against <about:blank>
  assert_equals: host expected "x:0" but got ""
[FAIL] Parsing: <chromeos-steam://x:0/a> against <about:blank>
  assert_equals: host expected "x:0" but got ""
[FAIL] Parsing: <steam://x:0/a> against <about:blank>
  assert_equals: host expected "x:0" but got ""
[FAIL] Parsing: <materialized-view://x:0/a> against <about:blank>
  assert_equals: host expected "x:0" but got ""
[FAIL] Parsing: <http://example.com/U+d800𐟾U+dfff﷐﷏﷯ﷰ￾￿?U+d800𐟾U+dfff﷐﷏﷯ﷰ￾￿> against <about:blank>
  assert_equals: href expected "http://example.com/%EF%BF%BD%F0%90%9F%BE%EF%BF%BD%EF%B7%90%EF%B7%8F%EF%B7%AF%EF%B7%B0%EF%BF%BE%EF%BF%BF?%EF%BF%BD%F0%90%9F%BE%EF%BF%BD%EF%B7%90%EF%B7%8F%EF%B7%AF%EF%B7%B0%EF%BF%BE%EF%BF%BF" but got "http://example.com/U+d800𐟾U+dfff﷐﷏﷯ﷰ\\ufffe\\uffff?U+d800𐟾U+dfff﷐﷏﷯ﷰ\\ufffe\\uffff"
Harness: the test ran to completion.
